Updated libraries to latest and fixed a bug

+++ src/Graphics/Renderer.cs
@@ -157,7 +157,7 @@ public class Renderer : MoonTools.ECS.Renderer {
                         str,
                         text.Size,
                         Matrix4x4.CreateTranslation(dropShadowPosition.X, dropShadowPosition.Y, depth - 1),
-                        new Color(0, 0, 0, color.A),
+                        new Color((float)0, 0, 0, color.A),
                         text.HorizontalAlignment,
                         text.VerticalAlignment
                     );
